introduce utils module
